Move get_env_short to ef_conf_utils
